fix: enforce localhost service defaults cross-platform (fixes #74)
