+++ cmd/tabura/main_test.go
@@ -1,0 +1,20 @@
+package main
+
+import "testing"
+
+func TestParseServerConfigDefaultsToLoopbackWebHost(t *testing.T) {
+	cfg, status := parseServerConfig([]string{})
+	if status != 0 {
+		t.Fatalf("parseServerConfig() status = %d, want 0", status)
+	}
+	if cfg.webHost != "127.0.0.1" {
+		t.Fatalf("default webHost = %q, want 127.0.0.1", cfg.webHost)
+	}
+}
+
+func TestParseServerConfigRejectsPublicMCPWithoutUnsafeFlag(t *testing.T) {
+	_, status := parseServerConfig([]string{"--mcp-host", "0.0.0.0"})
+	if status != 2 {
+		t.Fatalf("parseServerConfig(public mcp) status = %d, want 2", status)
+	}
+}
