+++ deploy/launchd/io.tabura.piper-tts.plist
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+  <key>Label</key>
+  <string>io.tabura.piper-tts</string>
+  <key>ProgramArguments</key>
+  <array>
+    <string>$HOME/.local/share/tabura-piper-tts/venv/bin/uvicorn</string>
+    <string>piper_tts_server:app</string>
+    <string>--app-dir</string>
+    <string>$HOME/code/assi/tabula/scripts</string>
+    <string>--host</string>
+    <string>127.0.0.1</string>
+    <string>--port</string>
+    <string>8424</string>
+  </array>
+  <key>EnvironmentVariables</key>
+  <dict>
+    <key>PIPER_MODEL_DIR</key>
+    <string>$HOME/.local/share/tabura-piper-tts/models</string>
+  </dict>
+  <key>RunAtLoad</key>
+  <true/>
+  <key>KeepAlive</key>
+  <true/>
+  <key>StandardOutPath</key>
+  <string>/tmp/tabura-piper-tts.log</string>
+  <key>StandardErrorPath</key>
+  <string>/tmp/tabura-piper-tts.log</string>
+</dict>
+</plist>
