+++ deploy/launchd/io.tabura.web.plist
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+  <key>Label</key>
+  <string>io.tabura.web</string>
+  <key>ProgramArguments</key>
+  <array>
+    <string>$HOME/.local/bin/tabura</string>
+    <string>server</string>
+    <string>--project-dir</string>
+    <string>$HOME/code/assi/tabula</string>
+    <string>--data-dir</string>
+    <string>$HOME/.local/share/tabura-web</string>
+    <string>--web-host</string>
+    <string>127.0.0.1</string>
+    <string>--web-port</string>
+    <string>8420</string>
+    <string>--mcp-host</string>
+    <string>127.0.0.1</string>
+    <string>--mcp-port</string>
+    <string>9420</string>
+    <string>--app-server-url</string>
+    <string>ws://127.0.0.1:8787</string>
+    <string>--tts-url</string>
+    <string>http://127.0.0.1:8424</string>
+  </array>
+  <key>RunAtLoad</key>
+  <true/>
+  <key>KeepAlive</key>
+  <true/>
+  <key>StandardOutPath</key>
+  <string>/tmp/tabura-web.log</string>
+  <key>StandardErrorPath</key>
+  <string>/tmp/tabura-web.log</string>
+</dict>
+</plist>
