+++ deploy/systemd/user/tabura-web.service
@@ -10,7 +10,7 @@ PassEnvironment=DISPLAY WAYLAND_DISPLAY XAUTHORITY XDG_RUNTIME_DIR DBUS_SESSION_
 Environment=HOME=%h
 Environment=SHELL=/usr/bin/bash
 Environment=TABURA_APP_SERVER_MODEL=gpt-5.3-codex-spark
-ExecStart=/usr/bin/bash -lc 'source ~/.bashrc >/dev/null 2>&1 || true; cd %h/code/assi/tabula; exec /usr/bin/env go run ./cmd/tabura server --project-dir %h/code/assi/tabula --data-dir %h/.local/share/tabura-web --web-host 0.0.0.0 --web-port 8420 --mcp-host 127.0.0.1 --mcp-port 9420 --app-server-url ws://127.0.0.1:8787 --tts-url http://127.0.0.1:8424 --dev-runtime'
+ExecStart=/usr/bin/bash -lc 'source ~/.bashrc >/dev/null 2>&1 || true; cd %h/code/assi/tabula; exec /usr/bin/env go run ./cmd/tabura server --project-dir %h/code/assi/tabula --data-dir %h/.local/share/tabura-web --web-host 127.0.0.1 --web-port 8420 --mcp-host 127.0.0.1 --mcp-port 9420 --app-server-url ws://127.0.0.1:8787 --tts-url http://127.0.0.1:8424 --dev-runtime'
 Restart=on-failure
 RestartSec=2
 TimeoutStopSec=5
